Joski_Games/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Deprecated==1.2.15
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
Markdown==3.7
multidict==6.1.0
nexmo==2.5.2
packaging==24.2
phonenumbers==8.13.51
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==75.6.0
sqlparse==0.5.2
twilio==9.3.8
urllib3==2.2.3
wrapt==1.17.0
yarl==1.18.3
